Split up Component tests
